Refactor MOA calculation logic into separate modules for improved organization and maintainability
